requirements.txt
numpy~=1.26.4
pandas~=2.2.2
scikit-learn~=1.5.1
tensorflow~=2.18.0
torch~=2.5.1
torchvision~=0.20.1
pillow~=11.0.0
tqdm